test: keep fixture helpers out of production path
The audit found two places where performance-oriented test changes widened risk: a production-exported prepared DB opener and a parallel test that mutates watcher timing globals. Both make the large fixture-sharing PR harder to reason about.

Keep the prepared opener test-only inside internal/db and use normal opens from cross-package helpers, and serialize the watcher timing test so global timing overrides cannot affect unrelated parallel tests.
